abstract optimisation loop
